+++ DataLayers/Layers/Coverage/SprinklerLayer.cs
@@ -43,8 +43,9 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
         /// <param name="translations">Provides translations in stored in the mod folder's i18n folder.</param>
         /// <param name="config">The data layer settings.</param>
         /// <param name="mods">Handles access to the supported mod integrations.</param>
-        public SprinklerLayer(ITranslationHelper translations, LayerConfig config, ModIntegrations mods)
-            : base(translations.Get("sprinklers.name"), config)
+        /// <param name="monitor">Writes messages to the SMAPI log.</param>
+        public SprinklerLayer(ITranslationHelper translations, LayerConfig config, ModIntegrations mods, IMonitor monitor)
+            : base(translations.Get("sprinklers.name"), config, monitor)
         {
             // init
             this.Mods = mods;