+++ DataLayers/Layers/TillableLayer.cs
@@ -37,8 +37,9 @@ namespace Pathoschild.Stardew.DataLayers.Layers
         /// <summary>Construct an instance.</summary>
         /// <param name="translations">Provides translations in stored in the mod folder's i18n folder.</param>
         /// <param name="config">The data layer settings.</param>
-        public TillableLayer(ITranslationHelper translations, LayerConfig config)
-            : base(translations.Get("tillable.name"), config)
+        /// <param name="monitor">Writes messages to the SMAPI log.</param>
+        public TillableLayer(ITranslationHelper translations, LayerConfig config, IMonitor monitor)
+            : base(translations.Get("tillable.name"), config, monitor)
         {
             this.Legend = new[]
             {